AD-95 Separated ModelDiffExtentions into its own module
